+++ src/main/java/com/servicekerdit/model/DetalleSunat.java
@@ -2,6 +2,7 @@ package com.servicekerdit.model;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 
 
 @Data
@@ -20,30 +21,30 @@ public class DetalleSunat
     private String codigo_producto_interno;
     private String codigo_producto_sunat;
     private String codigo_producto_gs1;
-    private double valor_unitario;
-    private double valor_venta;
-    private double descuento;
+    private BigDecimal valor_unitario;
+    private BigDecimal valor_venta;
+    private BigDecimal descuento;
     private String codigo_descuento;
-    private double precio_venta_unitario;
-    private double valor_referencial_unitario;
-    private double monto_base_igv;
-    private double monto_base_ivap;
-    private double monto_base_exportacion;
-    private double monto_base_exonerado;
-    private double monto_base_inafecto;
-    private double monto_base_gratuito;
-    private double monto_base_isc;
-    private double monto_base_otros_trib;
-    private double igv;
-    private double isc;
-    private double ivap;
-    private double imp_vta_grat;
-    private double imp_otros_trib;
-    private double porcentaje_igv;
-    private double procentaje_ivap;
-    private double porcentaje_isc;
-    private double porcentaje_otros_trib;
-    private double porcentaje_trib_vta_grat;
+    private BigDecimal precio_venta_unitario;
+    private BigDecimal valor_referencial_unitario;
+    private BigDecimal monto_base_igv;
+    private BigDecimal monto_base_ivap;
+    private BigDecimal monto_base_exportacion;
+    private BigDecimal monto_base_exonerado;
+    private BigDecimal monto_base_inafecto;
+    private BigDecimal monto_base_gratuito;
+    private BigDecimal monto_base_isc;
+    private BigDecimal monto_base_otros_trib;
+    private BigDecimal igv;
+    private BigDecimal isc;
+    private BigDecimal ivap;
+    private BigDecimal imp_vta_grat;
+    private BigDecimal imp_otros_trib;
+    private int porcentaje_igv;
+    private BigDecimal procentaje_ivap;
+    private BigDecimal porcentaje_isc;
+    private BigDecimal porcentaje_otros_trib;
+    private BigDecimal porcentaje_trib_vta_grat;
     private String tipo_afectacion_igv;
     private String tipo_isc;
